+++ src/main/com/fulcrologic/fulcro/dom.cljs
@@ -6,8 +6,8 @@
     [clojure.spec.alpha :as s]
     [clojure.string :as str]
     [com.fulcrologic.fulcro.algorithms.misc :as util]
-    [cljsjs.react]
-    [cljsjs.react.dom]
+    ["react" :as react]
+    ["react-dom" :as react-dom]
     [goog.object :as gobj]
     [goog.dom :as gdom]
     [com.fulcrologic.fulcro.dom-common :as cdom]
@@ -29,7 +29,7 @@
   u ul unknown use var video view vkern wbr)
 
 (def ^{:private true} element-marker
-  (-> (js/React.createElement "div" nil)
+  (-> (react/createElement "div" nil)
     (gobj/get "$$typeof")))
 
 (defn element? "Returns true if the given arg is a react element."
@@ -53,19 +53,16 @@
 (defn render
   "Equivalent to React.render"
   [component el]
-  (js/ReactDOM.render component el))
-
-(defn render-to-str
-  "Equivalent to React.renderToString. NOTE: You must require cljsjs.react.dom.server to use this function."
-  [c]
-  (js/ReactDOMServer.renderToString c))
+  (react-dom/render component el))
 
 (defn node
-  "Returns the dom node associated with a component's React ref."
+  "Returns the dom node associated with a component's React ref.
+
+  DEPRECATED. Use React's function-based refs."
   ([component]
-   (js/ReactDOM.findDOMNode component))
+   (react-dom/findDOMNode component))
   ([component name]
-   (some-> (.-refs component) (gobj/get name) (js/ReactDOM.findDOMNode))))
+   (some-> (.-refs component) (gobj/get name) (react-dom/findDOMNode))))
 
 (defn create-element
   "Create a DOM element for which there exists no corresponding function.
@@ -74,10 +71,9 @@
   ([tag]
    (create-element tag nil))
   ([tag opts]
-   (js/React.createElement tag opts))
+   (react/createElement tag opts))
   ([tag opts & children]
-   (js/React.createElement tag opts children)))
-
+   (react/createElement tag opts children)))
 
 (defn convert-props
   "Given props, which can be nil, a js-obj or a clj map: returns a js object."
@@ -96,7 +92,7 @@
   "Used internally by the DOM element generation."
   [arr]
   {:pre [(array? arr)]}
-  (.apply js/React.createElement nil arr))
+  (.apply react/createElement nil arr))
 
 (defn- update-state
   "Updates the state of the wrapped input element."
@@ -126,9 +122,9 @@
                        (gobj/extend state props))
                      (gobj/remove state "inputRef")
                      state))
-                 (.apply js/React.Component this (js-arguments))))]
+                 (.apply react/Component this (js-arguments))))]
     (set! (.-displayName ctor) (str "wrapped-" element))
-    (goog.inherits ctor js/React.Component)
+    (goog.inherits ctor react/Component)
     (specify! (.-prototype ctor)
       Object
       (onChange [this event]
@@ -140,7 +136,7 @@
 
       (componentWillReceiveProps [this new-props]
         (let [state-value   (gobj/getValueByKeys this "state" "value")
-              this-node     (js/ReactDOM.findDOMNode this)
+              this-node     (react-dom/findDOMNode this)
               value-node    (if (is-form-element? this-node)
                               this-node
                               (gdom/findNode this-node #(is-form-element? %)))
@@ -155,8 +151,8 @@
             (update-state this new-props (gobj/get new-props "value")))))
 
       (render [this]
-        (js/React.createElement element (.-state this))))
-    (let [real-factory (js/React.createFactory ctor)]
+        (react/createElement element (.-state this))))
+    (let [real-factory (react/createFactory ctor)]
       (fn [props & children]
         (if-let [r (gobj/get props "ref")]
           (if (string? r)