+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+3.0.0-alpha-9
+-------------
+- Improved ident-optimized render, added support for :only-refresh
+- Added specs and docs strings
+- Fixed some naming where registry key should have been used
+- Fixed componentDidMount prev-props (failing to update)
+- Fixed tempid rewrites on mutation return from server
+- Changed transit stack. Updated dev guide for it.
+- Deprecated some names
+- Added SSR render-to-string support
+- Switched to ghostwheel 0.4 for better spec elision in production builds
+
 3.0.0-alpha-8
 -------------
 - Improved logging helpers